TeleService: Add support for call timer reset on CDMA MO call
In android original design, the duration of CDMA MO call is started from
the dial command sent, so it is not the real duration of the active time.

In this patch, a new message is registered to listen the event of the call
accepted, and then reset the call duration when the event happens.

Change-Id: Idae08bc17ad0044390ad8ebd34f7b4f29d35f7ed
CRs-Fixed: 2207419