Recalculate conference on RTT mode change
When there is RTT mode change there will not be any call state change triggered
to UI as there is no change in call state or call details but only the call
capability is updated. This gives incorrect data to InCallUI on confernceable
calls. Recalculate confernce on RTT mode change to update UI on conferenceable
calls.

Change-Id: I0064549321b14119968ad12472f9d11181b5b701
CRs-Fixed: 2452956